requirements.txt
asyncio~=3.4.3
gspread==5.10.0
openpyxl~=3.1.2
google-auth==2.21.0
aiogram~=2.25.1
magic-filter == 1.0.9
requests~=2.31.0
lxml
psutil~=5.9.5
certifi~=2023.5.7
python-dotenv~=1.0.0
asyncpg~=0.27.0
aiohttp~=3.8.4
pandas~=2.0.3
XlsxWriter~=3.1.2